+++ CHANGELOG.md
@@ -2962,3 +2962,4 @@
 - [2026-07-20 18:49] style(forms): harmonize light and dark theme palette for forms
 - [2026-07-20 18:58] fix(auth): prevent double tap duplicate action in auth
 - [2026-07-20 19:07] fix(storage): fix memory leak by cancelling storage subscription
+- [2026-07-20 19:30] feat(navigation): add async navigation stream provider